refactor: domain 패키지 생성 및 클래스들 위치 변경 - BonusNumber, Lotto, Money 패키지 위치 domain 내부로 변경
