+++ src/main/java/lotto/domain/Money.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.ErrorMessage;
 
 public class Money {
